fix: the bug of missing a day in determineCardColor

+++ app/src/main/java/com/example/justdoit/list/ListUtils.kt
@@ -2,14 +2,14 @@ package com.example.justdoit.list
 
 import com.example.justdoit.R
 
-fun determineCardColor(dueDate: Long?, done: Boolean): Int {
+fun determineCardColor(dueDate: Double?, done: Boolean): Int {
     var color = R.color.todoNotDue
 
     if (done) return R.color.todoDone
 
     if (dueDate != null) {
-        val now = System.currentTimeMillis()
-        val day = 1000 * 60 * 60 * 24
+        val now = System.currentTimeMillis().toDouble()
+        val day = (1000 * 60 * 60 * 24).toDouble()
         val daysUntilDue = (dueDate.minus(now)) / day
 
         color = when {
